Linter findings

+++ x/emissions/keeper/keeper.go
@@ -792,7 +792,7 @@ func (k *Keeper) AppendInference(ctx context.Context, topicId TopicId, nonce typ
 		}
 	}
 	// append inference if not reached out topN
-	if len(newInferences.Inferences) < int(moduleParams.MaxTopInferersToReward) {
+	if uint64(len(newInferences.Inferences)) < moduleParams.MaxTopInferersToReward {
 		newInferences.Inferences = append(newInferences.Inferences, inference)
 		return k.allInferences.Set(ctx, key, newInferences)
 	}
@@ -848,7 +848,7 @@ func (k *Keeper) AppendForecast(ctx context.Context, topicId TopicId, nonce type
 			newForecasts.Forecasts = append(newForecasts.Forecasts, exForecast)
 		}
 	}
-	if len(newForecasts.Forecasts) < int(moduleParams.MaxTopForecastersToReward) {
+	if uint64(len(newForecasts.Forecasts)) < moduleParams.MaxTopForecastersToReward {
 		newForecasts.Forecasts = append(newForecasts.Forecasts, forecast)
 		return k.allForecasts.Set(ctx, key, newForecasts)
 	}
@@ -938,7 +938,7 @@ func (k *Keeper) AppendReputerLoss(ctx context.Context, topicId TopicId, block B
 			newReputerLossBundles.ReputerValueBundles = append(newReputerLossBundles.ReputerValueBundles, exReputation)
 		}
 	}
-	if len(newReputerLossBundles.ReputerValueBundles) < int(moduleParams.MaxTopReputersToReward) {
+	if uint64(len(newReputerLossBundles.ReputerValueBundles)) < moduleParams.MaxTopReputersToReward {
 		newReputerLossBundles.ReputerValueBundles = append(newReputerLossBundles.ReputerValueBundles, reputerLoss)
 		return k.allLossBundles.Set(ctx, key, newReputerLossBundles)
 	}
@@ -2286,7 +2286,7 @@ func (k *Keeper) GetInferenceScoresUntilBlock(ctx context.Context, topicId Topic
 	if err != nil {
 		return nil, err
 	}
-	maxNumTimeSteps := int(moduleParams.MaxSamplesToScaleScores)
+	maxNumTimeSteps := moduleParams.MaxSamplesToScaleScores
 
 	scores := make([]*types.Score, 0, maxNumTimeSteps)
 
@@ -2297,13 +2297,13 @@ func (k *Keeper) GetInferenceScoresUntilBlock(ctx context.Context, topicId Topic
 		}
 
 		for _, score := range existingScores.Value.Scores {
-			if len(scores) < maxNumTimeSteps {
+			if uint64(len(scores)) < maxNumTimeSteps {
 				scores = append(scores, score)
 			} else {
 				break
 			}
 		}
-		if len(scores) >= maxNumTimeSteps {
+		if uint64(len(scores)) >= maxNumTimeSteps {
 			break
 		}
 		iter.Next()
@@ -2363,7 +2363,7 @@ func (k *Keeper) GetForecastScoresUntilBlock(ctx context.Context, topicId TopicI
 	if err != nil {
 		return nil, err
 	}
-	maxNumTimeSteps := int(moduleParams.MaxSamplesToScaleScores)
+	maxNumTimeSteps := moduleParams.MaxSamplesToScaleScores
 
 	scores := make([]*types.Score, 0, maxNumTimeSteps)
 
@@ -2374,13 +2374,13 @@ func (k *Keeper) GetForecastScoresUntilBlock(ctx context.Context, topicId TopicI
 		}
 
 		for _, score := range existingScores.Value.Scores {
-			if len(scores) < maxNumTimeSteps {
+			if uint64(len(scores)) < maxNumTimeSteps {
 				scores = append(scores, score)
 			} else {
 				break
 			}
 		}
-		if len(scores) >= maxNumTimeSteps {
+		if uint64(len(scores)) >= maxNumTimeSteps {
 			break
 		}
 		iter.Next()
